requirements.txt
py-kelctl
pyserial
matplotlib